switch off the original stacks / axes draw (in draw-map).
move from draw-map.js to utils/draw/axisBrush.js : setupBrushZoom(), brushClipSize().
draw-map.js : add draw_orig = false and use to disable stacks / axes rendering in draw().
use pathDataUtils in place of axisApi for pathUpdate() in : view-controls.js, axis-menu-actions.js, stacks-drag.js.
axis-draw.js : draw() : use setupBrushZoom().
path-info.js : provide this=pathInfo in : hidePathHoverToolTip(), toolTipMouseOut(), closeToolTip(), setupToolTipMouseHover().
