+++ frontend/app/components/draw-map.js
@@ -180,6 +180,8 @@ const dLog = console.debug;
 Object.filter = Object_filter;
 
 
+const draw_orig = false;
+
 
 /*----------------------------------------------------------------------------*/
 
@@ -1870,7 +1872,8 @@ export default Component.extend(Evented, {
         })
       ;
     }
-    
+
+    if (draw_orig) {
     // pathUpdate(undefined);
     stacks.log();
 
@@ -1961,6 +1964,7 @@ export default Component.extend(Evented, {
       });
       console.log('remnant', removedStacks.node());
     }
+    }
 
     // moved to ../utils/draw/axis.js : stackEltId()
 
@@ -1975,6 +1979,7 @@ export default Component.extend(Evented, {
       return result;
     }
 
+    if (draw_orig) {
     const
     axisDraw = new AxisDraw(oa, /*axis1d*/null, stacks, /*stacksView*/ null),
     selections = {svgContainer, stackSd, stackS,  stackX};    
@@ -1990,33 +1995,7 @@ export default Component.extend(Evented, {
 
 //- moved to ../utils/draw/axis.js : yAxisTextScale(),  yAxisTicksScale(),  yAxisBtnScale()
 
-    // Add a brush for each axis.
-    let gBrushParent =
-    allG.append("g")
-      .attr("class", "brush");
-    /** Ensure there is clipPath & rect in gBrushParent, and set its geometry. */
-    function brushClipSize(gBrushParent) {
-      gBrushParent
-      .each(function(axisID) {
-        brushClip(d3.select(this), axisID)
-          .each(function(d) { d3.select(this).call(oa.y[d].brush); });
-      });
-    }
-    /** brushClip() uses getBBox(), so call again after the geometry has settled.
-     * If this works reliably, it might suggest splitting the append(clipPath)
-     * from the geometry setting.
-     * The 2sec call should be fine for most computers, some might take until
-     * 5sec to settle the geometry.
-     */
-    brushClipSize(gBrushParent);
-    later(() => ! me.isDestroying && brushClipSize(gBrushParent), 2000);
-    later(() => ! me.isDestroying && brushClipSize(gBrushParent), 5000);
-
-
-    if (allG.nodes().length)
-      console.log('zoomBehavior', allG.nodes(), allG.node());
-    allG
-      .call(oa.zoomBehavior);
+      //- moved to utils/draw/axisBrush.js : setupBrushZoom(), brushClipSize()
 
     /*------------------------------------------------------------------------*/
     /* above is the setup of scales, stacks, axis */
@@ -2024,6 +2003,7 @@ export default Component.extend(Evented, {
     if (source == 'dataReceived')
       stacks.changed = 0x10;
     let t = stacksAdjust(true, undefined);
+    }
     /* below is the setup of path hover (path classes, colouring are setup
      * above, but that can be moved following this, when split out). */
     /*------------------------------------------------------------------------*/
